feat: color grid only when left mouse button down
Currently, end-users will start drawing on the grid as soon as their
mouse pointer enters the grid container and hovers over the grid
squares.

This change will give end-users the ability to control where
they would like to draw on the grid. End-users can either
color one grid square at a time, or they can color grid squares
the mouse pointer hovers over so long as the left button is
pressed down.

To prevent mistakes when drawing on the grid, no drawing
action performs when the end-user attempts to draw
on the grid while clicking on the mouse's right button.