fix: correct grid behavior on grid reset
When a end user cancels the reset prompt, the grid disappears
and leaves behind an empty grid container. This change ensures that the
current grid layout does not disappear should the end user decide to
cancel the reset grid prompt.

At this time active buttons do not disable if the end user
successfully resets the grid contaner with new grid
squares. This change sets any active buttons to
false at the time of a grid reset, while leaving
normalColoringBool set to true.

At this time the color selector is not resetting to #FFFFFF
when the grid successfully resets. This change will reset
the color selector to have #FFFFFF as its value when the
end user resets their grid layout.