feat: add first version of undo button
At this time, there is no option available to undo drawing
instances made on the grid. This is a disservice for the end
user as there is no way to undo any mistakes made while
coloring on the grid.

This first version of the undo button will allow end users to revert
their grid to only one previous drawing instance, that being the
most recent.