feat: allow end users to undo multiple times
At this time, end users can undo to only the most recent
drawing action. To improve the experience, I
would like to make it possible to let the end user
undo as far back as the 5 most recent drawing instances.

This is still not complete, but I wanted to save my work as I
have successfully created a way to log multiple
drawing instances previously performed inside the
historicalColoring array. There is no condition implemented
just yet that ensures only 5 drawing instances are stored at a
time. As a result, the historicalColoring array will store
all the drawing instances performed until the grid is reset.

A condition will be implemented to ensure that only 5
instances are saved at a time, and testing will be
performed to ensure that this new method will allow the
end user to perform Redo action without any
complications.