feat: undo and redo functionality second version
Currently, the undo button allows the end-user to reverse their
normal coloring action as far back as the beginning of the drawing
session. However, the current version does not store instances
where the end-user had changed the background color, cleared
the grid, and colored over the same square during a drawing
instance. The end user is also unable to redo instances at
this time.

This push allows the end-user to undo and redo instances
such as:

- Background Fill
- Clearing Grid
- Eraser Coloring
- Normal Coloring
- Darken Coloring
- Lighten Coloring
- Rainbow Coloring

There is no limit as to how far back the end-user can undo.

This commit includes repeated lines of code, therefore
refactoring will take place and a new version will be pushed
out in the near future.